Update README.md
 Thanks @xacrimon for pointing this out!

Context: On Reddit we discussed how even though it was stated that Lariv and DashMap were very different data structures, the comparison was a bit misleading. Even though the discussion was probably caused by the wording of the post's title, as it was not stated there, I want to fix it here too.